move nugetCommandArgs and consume_test to artifactory/utils/dotnet
